Change time cycle flag for GEOS-Chem restart file to ensure file for correct date is always supplied
The SPC_ fields from the GEOS-Chem restart file now use time cycle flag "EY"
for "exact, simulation year" to ensure that the restart file for the start
date is always provided. The previous "CYS" option would allow for cycling
if a restart file for a different data was supplied and for skipping if a
species wasn't found in the restart file, both of which may produce unintended
initial conditions.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>